fix(kotlin): HopStatus.forwardHops is UByte (F-1, the F-9 sibling) + a real reinterpretation test (F-2) (#99)
Pass-18's Android audit found F-9 (HopMessage.hops Byte -> UByte, PR #91) had a sibling it missed:
HopStatus.forwardHops was still `Byte`, populated from hop_message_status's uint8_t out_hops. The exact
same latent overflow, a forward-hop count >= 128 would render negative, just reached via status(id)
instead of pollInbox. Changed forwardHops to UByte (reinterpreted from the FFI Byte via .toUByte() at
the public surface; the FFI boundary keeps Byte for correct JNA marshalling of the native uint8_t).

F-2: the audit also flagged that F-9's own regression check `assertTrue(msg.hops >= 0u)` is now
tautological (always true for a UByte). Added uint8HopCountsSurfaceUnsignedNotNegative, which exercises
the ACTUAL bug the UByte fields exist to prevent: a native uint8_t of 200 marshals as the signed Byte
-56, and the public field must read 200u (not negative). Asserts this directly for both HopMessage.hops
and HopStatus.forwardHops at the >127 boundary a real mesh test never reaches.

Kotlin SDK tests + the 80% coverage gate BUILD SUCCESSFUL.


GitOrigin-RevId: e4169b29cc101d50aa19ba6922f919b93f9bc0ae

+++ sdk/wrappers/kotlin/src/main/kotlin/sh/hop/Hop.kt
@@ -57,8 +57,11 @@ data class HopStatus(
     val relayed: Int,
     /** Destination confirmed. */
     val delivered: Boolean,
-    /** Forward-path length the destination reported. */
-    val forwardHops: Byte,
+    /** Forward-path length the destination reported. `UByte`, not `Byte`: the C ABI field is uint8_t
+     *  (0..255), so a signed Byte would render a forward-hop count >= 128 as negative (pass-18 F-1, the
+     *  sibling of the HopMessage.hops F-9 fix). The FFI boundary keeps Byte for correct JNA marshalling;
+     *  we reinterpret to UByte at this public surface. */
+    val forwardHops: UByte,
     /** Forward-path latency (ms) the destination reported. */
     val forwardMs: Int,
 )
@@ -306,7 +309,7 @@ class HopNode private constructor(rawPtr: Pointer) : AutoCloseable {
         return HopStatus(
             relayed = relayed.value,
             delivered = delivered.value.toInt() != 0,
-            forwardHops = hops.value,
+            forwardHops = hops.value.toUByte(),
             forwardMs = ms.value,
         )
     }
